+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2] - 2023-08-15
+
+### Added
+- MultiChannel join functionallity
+    - Join Public Method to join a second channel
+    - Chatter.IsFromChannel to check if a message is in one of the channels
+
+### Changed
+- Namespace is now called VerySimpleTwitchChat in order to integrate it with other Twitch plugins
+- Classes renamed
+    - TwitchControllerEditor -> TwitchChatEditor
+    - TwitchController -> TwitchChat
+    - TwitchSettingsEditor -> TwitchChatSettingsEditor
+    - TwitchSettings -> TwitchChatSettings
+    - TwitchSettingsWindowEditor -> TwitchChatSettingsWindowEditor
+
 ## [1.1.1] - 2023-07-18
 
 ### Added